tests for Transshipment
