+++ test/Transshipment/index.test.ts
@@ -1,3 +1,6 @@
-// describe("Transshipment: ", () => {
-//     require("./constructor.test");
-//   });
+describe("Transshipment: ", () => {
+  require("./constructor.test");
+  require("./createAcc.test");
+  require("./sendMassages.test");
+  require("./sendUniversalMassage.test");
+});
