+++ test/Transshipment/sendMassages.test.ts
@@ -1,0 +1,336 @@
+import { expect } from "chai";
+import { BigNumber, ContractTransaction } from "ethers";
+import { ethers, getChainId } from "hardhat";
+import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
+import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
+import {
+  sign,
+  Domain,
+  typesForBridge,
+  ZERO_ADDRESS,
+  ZERO_BYTES,
+  standardPrepare,
+  MassageParamStructAbi,
+  BridgeParamsStruct,
+} from "@test-utils";
+import { ITransshipmentStructures } from "@contracts/Transshipment";
+import { ParamType } from "@ethersproject/abi";
+
+describe("Method: sendMassage: ", () => {
+  describe("When one of parameters is incorrect", () => {
+    it("When try check", () => {
+      expect(true);
+    });
+  });
+
+  describe("When all parameters correct ", () => {
+    let result: ContractTransaction;
+
+    it("should success send tokens from EOA (fee in token)", async () => {
+      const { link, srcUSDC, transshipmentSender, transshipmentReceiver, user } = await loadFixture(
+        standardPrepare
+      );
+
+      const massageParam: ITransshipmentStructures.MassageParamStruct = {
+        destinationChainSelector: 2,
+        receiver: transshipmentReceiver.address, // address at bsc
+        dataToSend: ZERO_BYTES,
+        addressToExecute: ZERO_ADDRESS,
+        valueToExecute: 0,
+        dataToExecute: ZERO_BYTES,
+        token: srcUSDC.address,
+        amount: ethers.utils.parseUnits("10", 18),
+        feeToken: link.address,
+        gasLimit: 200000,
+      };
+
+      await link.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      await srcUSDC.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await srcUSDC.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      result = await transshipmentSender.connect(user).sendMassage(massageParam);
+
+      await expect(result).to.be.not.reverted;
+      console.log(await srcUSDC.balanceOf(transshipmentReceiver.address));
+      console.log(await srcUSDC.balanceOf(user.address));
+    });
+
+    it("should success send tokens from EOA (fee in ETH)", async () => {
+      const { link, srcUSDC, transshipmentSender, transshipmentReceiver, user } = await loadFixture(
+        standardPrepare
+      );
+
+      const massageParam: ITransshipmentStructures.MassageParamStruct = {
+        destinationChainSelector: 2,
+        receiver: transshipmentReceiver.address, // address at bsc
+        dataToSend: ZERO_BYTES,
+        addressToExecute: ZERO_ADDRESS,
+        valueToExecute: 0,
+        dataToExecute: ZERO_BYTES,
+        token: srcUSDC.address,
+        amount: ethers.utils.parseUnits("10", 18),
+        feeToken: ZERO_ADDRESS,
+        gasLimit: 200000,
+      };
+
+      await srcUSDC.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await srcUSDC.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      result = await transshipmentSender
+        .connect(user)
+        .sendMassage(massageParam, { value: ethers.utils.parseUnits("1", 18) });
+
+      await expect(result).to.be.not.reverted;
+      console.log(await srcUSDC.balanceOf(transshipmentReceiver.address));
+      console.log(await srcUSDC.balanceOf(user.address));
+    });
+
+    it("should success send ETH from EOA (fee in ETH)", async () => {
+      const { link, srcUSDC, transshipmentSender, transshipmentReceiver, user } = await loadFixture(
+        standardPrepare
+      );
+      let contractBalance = await ethers.provider.getBalance(user.address);
+      console.log(contractBalance);
+      const massageParam: ITransshipmentStructures.MassageParamStruct = {
+        destinationChainSelector: 2,
+        receiver: transshipmentReceiver.address, // address at bsc
+        dataToSend: ZERO_BYTES,
+        addressToExecute: ZERO_ADDRESS,
+        valueToExecute: 0,
+        dataToExecute: ZERO_BYTES,
+        token: ZERO_ADDRESS,
+        amount: 0,
+        feeToken: ZERO_ADDRESS,
+        gasLimit: 200000,
+      };
+
+      result = await transshipmentSender
+        .connect(user)
+        .sendMassage(massageParam, { value: ethers.utils.parseUnits("5", 18) });
+
+      contractBalance = await ethers.provider.getBalance(transshipmentSender.address);
+      console.log(contractBalance);
+
+      await expect(result).to.be.not.reverted;
+    });
+
+    it("should success send ETH from EOA (fee in token)", async () => {
+      const { link, srcUSDC, transshipmentSender, transshipmentReceiver, user } = await loadFixture(
+        standardPrepare
+      );
+      let contractBalance = await ethers.provider.getBalance(user.address);
+      console.log(contractBalance);
+      const massageParam: ITransshipmentStructures.MassageParamStruct = {
+        destinationChainSelector: 2,
+        receiver: transshipmentReceiver.address, // address at bsc
+        dataToSend: ZERO_BYTES,
+        addressToExecute: ZERO_ADDRESS,
+        valueToExecute: 0,
+        dataToExecute: ZERO_BYTES,
+        token: ZERO_ADDRESS,
+        amount: 0,
+        feeToken: link.address,
+        gasLimit: 200000,
+      };
+
+      await link.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      result = await transshipmentSender
+        .connect(user)
+        .sendMassage(massageParam, { value: ethers.utils.parseUnits("5", 18) });
+
+      contractBalance = await ethers.provider.getBalance(transshipmentSender.address);
+      console.log(contractBalance);
+
+      await expect(result).to.be.not.reverted;
+    });
+
+    it("should success send massage from account to account for transfer tokens", async () => {
+      const { link, srcUSDC, dstUSDC, transshipmentSender, transshipmentReceiver, manager, user } =
+        await loadFixture(standardPrepare);
+
+      const userSrcAccountAddress = await transshipmentSender.getAccountAddress(user.address);
+      const userDstAccountAddress = await transshipmentReceiver.getAccountAddress(user.address);
+
+      await link.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      await srcUSDC.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await srcUSDC.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      await transshipmentSender.connect(user).createAccount("name", 1);
+      const userSrcAccount = (await ethers.getContractFactory("Account")).attach(userSrcAccountAddress);
+
+      await transshipmentReceiver.connect(user).createAccount("name", 1);
+      const userDstAccount = (await ethers.getContractFactory("Account")).attach(userDstAccountAddress);
+      await dstUSDC.connect(user).mint(userDstAccount.address, 1000);
+
+      const encodedTransfer = dstUSDC.interface.encodeFunctionData("transfer", [manager.address, 100]); // Encode transfer from dstAcc to manager
+      const encodedData = userDstAccount.interface.encodeFunctionData("execute", [
+        dstUSDC.address,
+        0,
+        encodedTransfer,
+      ]); // Encode call user dst account call usdc for transfer
+
+      const dstMassageParam: ITransshipmentStructures.MassageParamStruct = {
+        destinationChainSelector: 0,
+        receiver: ZERO_ADDRESS,
+        dataToSend: ZERO_BYTES,
+        addressToExecute: userDstAccount.address,
+        valueToExecute: 0,
+        dataToExecute: encodedData,
+        token: ZERO_ADDRESS,
+        amount: 0,
+        feeToken: ZERO_ADDRESS,
+        gasLimit: 0,
+      };
+
+      const encodedDstMassageParam = ethers.utils.defaultAbiCoder.encode(
+        MassageParamStructAbi as ParamType[],
+        [dstMassageParam]
+      );
+
+      const massageParam: ITransshipmentStructures.MassageParamStruct = {
+        destinationChainSelector: "2",
+        receiver: transshipmentReceiver.address,
+        dataToSend: encodedDstMassageParam,
+        addressToExecute: ZERO_ADDRESS,
+        valueToExecute: 0,
+        dataToExecute: ZERO_BYTES,
+        token: ZERO_ADDRESS,
+        amount: 0,
+        feeToken: ZERO_ADDRESS,
+        gasLimit: 200000,
+      };
+
+      const fees = ethers.utils.parseUnits("1", 18);
+
+      const encodedTransshipmentCallData = transshipmentReceiver.interface.encodeFunctionData("sendMassage", [
+        massageParam,
+      ]);
+
+      result = await userSrcAccount
+        .connect(user)
+        .execute(transshipmentSender.address, fees, encodedTransshipmentCallData, { value: fees });
+
+      console.log(await dstUSDC.balanceOf(userDstAccount.address));
+
+      await expect(result).to.be.not.reverted;
+    });
+
+    it("should success bridge tokens for EOA with account bridge (fee in ETH)", async () => {
+      const { link, srcUSDC, dstUSDC, transshipmentSender, manager, user, alice, srcDomain } =
+        await loadFixture(standardPrepare);
+
+      const userSrcAccountAddress = await transshipmentSender.getAccountAddress(user.address);
+      const userDstAccountAddress = userSrcAccountAddress; // one address for all networks
+      console.count();
+
+      await link.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      // await srcUSDC.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      // await srcUSDC.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      await srcUSDC.connect(alice).mint(alice.address, ethers.utils.parseUnits("100", 18));
+      await srcUSDC.connect(alice).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      await transshipmentSender.connect(user).createAccount("name_src_1", 1);
+      const userSrcAccount = (await ethers.getContractFactory("Account")).attach(userSrcAccountAddress);
+
+      // await transshipmentReceiver.connect(user).createAccount("name_dst_1", 1);
+      const userDstAccount = (await ethers.getContractFactory("Account")).attach(userDstAccountAddress);
+      await dstUSDC.connect(user).mint(userDstAccount.address, 1000);
+
+      console.count();
+
+      const bridgeParams: BridgeParamsStruct = {
+        userAddress: alice.address,
+        userNonce: await transshipmentSender.userNonce(alice.address),
+        srcTokenAddress: srcUSDC.address,
+        srcTokenAmount: 110,
+        dstChainSelector: 2,
+        dstExecutor: userDstAccount.address, // eq to srcAccount
+        dstTokenAddress: dstUSDC.address,
+        dstTokenAmount: 100,
+        dstReceiver: alice.address,
+      };
+
+      console.count();
+
+      const managerSignature = await sign(srcDomain, typesForBridge, bridgeParams, manager);
+
+      console.log("managerSignature: ", managerSignature);
+
+      const fees = ethers.utils.parseUnits("1", 18);
+
+      result = await transshipmentSender
+        .connect(alice)
+        .bridgeTokens(managerSignature, ZERO_ADDRESS, 200000, fees, bridgeParams, { value: fees });
+
+      console.log(await dstUSDC.balanceOf(alice.address));
+
+      await expect(result).to.be.not.reverted;
+    });
+
+    it("should success bridge tokens for EOA with account bridge (fee in Token)", async () => {
+      const { link, srcUSDC, dstUSDC, transshipmentSender, manager, user, alice, srcDomain } =
+        await loadFixture(standardPrepare);
+
+      const userSrcAccountAddress = await transshipmentSender.getAccountAddress(user.address);
+      const userDstAccountAddress = userSrcAccountAddress; // one address for all networks
+      console.count();
+
+      await link.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      // await srcUSDC.connect(user).mint(user.address, ethers.utils.parseUnits("100", 18));
+      // await srcUSDC.connect(user).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(alice).mint(alice.address, ethers.utils.parseUnits("100", 18));
+      await link.connect(alice).approve(transshipmentSender.address, ethers.utils.parseUnits("1", 18));
+
+      await srcUSDC.connect(alice).mint(alice.address, ethers.utils.parseUnits("100", 18));
+      await srcUSDC.connect(alice).approve(transshipmentSender.address, ethers.utils.parseUnits("100", 18));
+
+      await transshipmentSender.connect(user).createAccount("name_src_1", 1);
+      const userSrcAccount = (await ethers.getContractFactory("Account")).attach(userSrcAccountAddress);
+
+      // await transshipmentReceiver.connect(user).createAccount("name_dst_1", 1);
+      const userDstAccount = (await ethers.getContractFactory("Account")).attach(userDstAccountAddress);
+      await dstUSDC.connect(user).mint(userDstAccount.address, 1000);
+
+      console.count();
+
+      const bridgeParams: BridgeParamsStruct = {
+        userAddress: alice.address,
+        userNonce: await transshipmentSender.userNonce(alice.address),
+        srcTokenAddress: srcUSDC.address,
+        srcTokenAmount: 110,
+        dstChainSelector: 2,
+        dstExecutor: userDstAccount.address, // eq to srcAccount
+        dstTokenAddress: dstUSDC.address,
+        dstTokenAmount: 100,
+        dstReceiver: alice.address,
+      };
+
+      console.count();
+
+      const managerSignature = await sign(srcDomain, typesForBridge, bridgeParams, manager);
+
+      console.log("managerSignature: ", managerSignature);
+
+      const fees = ethers.utils.parseUnits("1", 18);
+
+      result = await transshipmentSender
+        .connect(alice)
+        .bridgeTokens(managerSignature, link.address, 200000, fees, bridgeParams);
+
+      console.log(await dstUSDC.balanceOf(alice.address));
+
+      await expect(result).to.be.not.reverted;
+    });
+  });
+});
