+++ test/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./constants";
 export * from "./signer";
 export * from "./accounts";
 export * from "./signature";
+export * from "./helpers";
